Make use of tokens for REST2 authentication
RT 5.0 has integrated v2 of the REST API and the Token authentication
module into core.

Tokens are used specifically for accessing the API.

Without this directive, the tokens are not permitting access to the API.

ps. I've left out the apache.rt.conf file in RT 4.x and the template
since I don't believe the Token authentication plugin is installed in
the 4.x images.